Elimination of custom code now provided by the standard library. Introduction of GSL. Performance improvements.

+++ src/VulkanEngine.h
@@ -746,7 +746,7 @@ namespace hg {
             submitInfo.pWaitSemaphores = waitSemaphores;
             submitInfo.pWaitDstStageMask = waitStages;
 
-            submitInfo.commandBufferCount = renderedCommandBuffers.size();
+            submitInfo.commandBufferCount = gsl::narrow<uint32_t>(renderedCommandBuffers.size());
             submitInfo.pCommandBuffers = renderedCommandBuffers.data();
 
             VkSemaphore signalSemaphores[] = { renderFinishedSemaphores[currentFrame].semaphore };
